sorted notes logic into composable useNotes.js
